Refactor: renamed functions and files related to atomic PDEVS models to start preparing for a second formalism implementation (DEVS)
