+++ example/main-count-fives.cpp
@@ -34,9 +34,9 @@
 #include <cadmium/modeling/ports.hpp>
 #include <cadmium/concept/coupled_model_assert.hpp>
 #include <cadmium/engine/pdevs_runner.hpp>
-#include <cadmium/basic_model/accumulator.hpp>
-#include <cadmium/basic_model/int_generator_one_sec.hpp>
-#include <cadmium/basic_model/reset_generator_five_sec.hpp>
+#include <cadmium/basic_model/pdevs/accumulator.hpp>
+#include <cadmium/basic_model/pdevs/int_generator_one_sec.hpp>
+#include <cadmium/basic_model/pdevs/reset_generator_five_sec.hpp>
 #include <cadmium/logger/common_loggers.hpp>
 using namespace std;
