+++ include/cadmium/modeling/dynamic_atomic.hpp
@@ -71,7 +71,7 @@ namespace cadmium {
 
                 atomic() {
                     static_assert(cadmium::concept::is_atomic<ATOMIC>::value, "This is not an atomic model");
-                    cadmium::concept::atomic_model_assert<ATOMIC>();
+                    cadmium::concept::pdevs_atomic_model_assert<ATOMIC>();
                     _id = boost::typeindex::type_id<model_type>().pretty_name();
                     _input_ports = cadmium::dynamic::modeling::create_dynamic_ports<input_ports>();
                     _output_ports = cadmium::dynamic::modeling::create_dynamic_ports<output_ports>();
@@ -79,7 +79,7 @@ namespace cadmium {
 
                 atomic(const std::string& model_id, Args&&... args) : ATOMIC<TIME>(std::forward<Args>(args)...) {
                     static_assert(cadmium::concept::is_atomic<ATOMIC>::value, "This is not an atomic model");
-                    cadmium::concept::atomic_model_assert<ATOMIC>();
+                    cadmium::concept::pdevs_atomic_model_assert<ATOMIC>();
                     _id = model_id;
                     _input_ports = cadmium::dynamic::modeling::create_dynamic_ports<input_ports>();
                     _output_ports = cadmium::dynamic::modeling::create_dynamic_ports<output_ports>();
