+++ test-compile/compile-fails/pdevs_missing_confluence_transition_fails_compile_test.cpp
@@ -25,37 +25,35 @@
  */
 
 /**
- * Test that when output ports of an atomic model are not defined as a tuple, atomic_model_assert fails compilation
+ * Test that an atomic model with no confluence transition fails compilation on atomic_model_assert
  */
 
 #include<cadmium/modeling/ports.hpp>
 #include<cadmium/concept/atomic_model_assert.hpp>
 #include<tuple>
 #include<cadmium/modeling/message_bag.hpp>
-#include<vector>
 
 /**
  * This model has no logic, only used for structural validation tests
  */
 template<typename TIME>
-struct atomic_model_with_outputs_as_vector
+struct atomic_model_missing_confluence_function
 {
     struct in : public cadmium::in_port<int>{};
     struct out : public cadmium::out_port<int>{};
 
-    constexpr atomic_model_with_outputs_as_vector() noexcept {}
+    constexpr atomic_model_missing_confluence_function() noexcept {}
     using state_type=int;
     state_type state=0;
     using input_ports=std::tuple<in>;
-    using output_ports=std::vector<out>;
+    using output_ports=std::tuple<out>;
 
     void internal_transition(){}
     void external_transition(TIME e, typename cadmium::make_message_bags<input_ports>::type mbs){}
-    void confluence_transition(TIME e, typename cadmium::make_message_bags<input_ports>::type mbs){}
     typename cadmium::make_message_bags<output_ports>::type output() const{}
     TIME time_advance() const{}
 };
 
 int main(){
-    cadmium::concept::atomic_model_assert<atomic_model_with_outputs_as_vector>();
+    cadmium::concept::pdevs_atomic_model_assert<atomic_model_missing_confluence_function>();
 }
