+++ test-compile/compile-fails/pdevs_missing_external_transition_fails_compile_test.cpp
@@ -25,21 +25,35 @@
  */
 
 /**
- * Test that a simulator of a valid atomic model provides the model type
+ * Test that an atomic model with no external transition fails compilation on atomic_model_assert
  */
 
 #include<cadmium/modeling/ports.hpp>
 #include<cadmium/concept/atomic_model_assert.hpp>
 #include<tuple>
 #include<cadmium/modeling/message_bag.hpp>
-#include <cadmium/basic_model/accumulator.hpp>
-#include <cadmium/engine/pdevs_simulator.hpp>
 
+/**
+ * This model has no logic, only used for structural validation tests
+ */
 template<typename TIME>
-using int_accumulator=cadmium::basic_models::accumulator<int, TIME>;
+struct atomic_model_missing_external_function
+{
+    struct in : public cadmium::in_port<int>{};
+    struct out : public cadmium::out_port<int>{};
+
+    constexpr atomic_model_missing_external_function() noexcept {}
+    using state_type=int;
+    state_type state=0;
+    using input_ports=std::tuple<in>;
+    using output_ports=std::tuple<out>;
+
+    void internal_transition(){}
+    void confluence_transition(TIME e, typename cadmium::make_message_bags<input_ports>::type mbs){}
+    typename cadmium::make_message_bags<output_ports>::type output() const{}
+    TIME time_advance() const{}
+};
 
 int main(){
-    using simulation_t = cadmium::engine::simulator<int_accumulator, float, cadmium::logger::not_logger>;
-    using model_t=simulation_t::model_type;
-    model_t model;
+    cadmium::concept::pdevs_atomic_model_assert<atomic_model_missing_external_function>();
 }
