+++ test-compile/compile-fails/pdevs_missing_internal_transition_fails_compile_test.cpp
@@ -25,8 +25,7 @@
  */
 
 /**
-* Test that an atomic model with no output_ports transition fails compilation on atomic_model_assert
- * This is different to say that it has an empty tuple of ports, which is a valid model definition
+ * Test that an atomic model with no internal transition fails compilation on atomic_model_assert
  */
 
 #include<cadmium/modeling/ports.hpp>
@@ -35,27 +34,26 @@
 #include<cadmium/modeling/message_bag.hpp>
 
 /**
- * This model has no logic, only used for structural validation tests.
- * In this case it is missing the declaration of input_ports type
- * For external and confluence transition fuctions defined input as empty tuple of ports
+ * This model has no logic, only used for structural validation tests
  */
 template<typename TIME>
-struct atomic_model_with_no_output_ports
+struct atomic_model_with_no_internal_transition
 {
     struct in : public cadmium::in_port<int>{};
+    struct out : public cadmium::out_port<int>{};
 
-    constexpr atomic_model_with_no_output_ports() noexcept {}
+    constexpr atomic_model_with_no_internal_transition() noexcept {}
     using state_type=int;
     state_type state=0;
     using input_ports=std::tuple<in>;
+    using output_ports=std::tuple<out>;
 
-    void internal_transition(){}
     void external_transition(TIME e, typename cadmium::make_message_bags<input_ports>::type mbs){}
     void confluence_transition(TIME e, typename cadmium::make_message_bags<input_ports>::type mbs){}
-    typename cadmium::make_message_bags<std::tuple<>>::type output() const{}
+    typename cadmium::make_message_bags<output_ports>::type output() const{}
     TIME time_advance() const{}
 };
 
 int main(){
-    cadmium::concept::atomic_model_assert<atomic_model_with_no_output_ports>();
+    cadmium::concept::pdevs_atomic_model_assert<atomic_model_with_no_internal_transition>();
 }
