+++ test-compile/compile-fails/pdevs_missing_output_function_fails_compile_test.cpp
@@ -25,16 +25,36 @@
  */
 
 /**
- * Test that passive, when used as expected create valid atomic models not failing compilation on atomic_model_assert.
+ * Test that an atomic model with no output function fails compilation on atomic_model_assert
  */
 
+#include<cadmium/modeling/ports.hpp>
 #include<cadmium/concept/atomic_model_assert.hpp>
-#include<cadmium/basic_model/passive.hpp>
+#include<tuple>
+#include<cadmium/modeling/message_bag.hpp>
 
-//preparing the passive to be used as atomic model
+/**
+ * This model has no logic, only used for structural validation tests
+ */
 template<typename TIME>
-using floating_passive=cadmium::basic_models::passive<float, TIME>;
+struct atomic_model_missing_output_function
+{
+    struct in : public cadmium::in_port<int>{};
+    struct out : public cadmium::out_port<int>{};
+
+    constexpr atomic_model_missing_output_function() noexcept {}
+    using state_type=int;
+    state_type state=0;
+    using input_ports=std::tuple<in>;
+    using output_ports=std::tuple<out>;
+
+    void internal_transition(){}
+    void external_transition(TIME e, typename cadmium::make_message_bags<input_ports>::type mbs){}
+    void confluence_transition(TIME e, typename cadmium::make_message_bags<input_ports>::type mbs){}
+
+    TIME time_advance() const{}
+};
 
 int main(){
-    cadmium::concept::atomic_model_assert<floating_passive>();
+    cadmium::concept::pdevs_atomic_model_assert<atomic_model_missing_output_function>();
 }
