+++ test-compile/compile-fails/pdevs_missing_output_ports_fails_compile_test.cpp
@@ -25,7 +25,8 @@
  */
 
 /**
- * Test that a valid atomic model does not stop compilation on atomic_model_assert.
+* Test that an atomic model with no output_ports transition fails compilation on atomic_model_assert
+ * This is different to say that it has an empty tuple of ports, which is a valid model definition
  */
 
 #include<cadmium/modeling/ports.hpp>
@@ -34,28 +35,27 @@
 #include<cadmium/modeling/message_bag.hpp>
 
 /**
-* Test that when an atomic model has duplicated output ports, atomic_model_assert fails compilation */
+ * This model has no logic, only used for structural validation tests.
+ * In this case it is missing the declaration of input_ports type
+ * For external and confluence transition fuctions defined input as empty tuple of ports
+ */
 template<typename TIME>
-struct atomic_model_with_repeated_output_ports
+struct atomic_model_with_no_output_ports
 {
     struct in : public cadmium::in_port<int>{};
 
-    struct out_one : public cadmium::out_port<int>{};
-    struct out_two : public cadmium::out_port<int>{};
-
-    constexpr atomic_model_with_repeated_output_ports() noexcept {}
+    constexpr atomic_model_with_no_output_ports() noexcept {}
     using state_type=int;
     state_type state=0;
     using input_ports=std::tuple<in>;
-    using output_ports=std::tuple<out_one, out_two, out_one>;
 
     void internal_transition(){}
     void external_transition(TIME e, typename cadmium::make_message_bags<input_ports>::type mbs){}
     void confluence_transition(TIME e, typename cadmium::make_message_bags<input_ports>::type mbs){}
-    typename cadmium::make_message_bags<output_ports>::type output() const{}
+    typename cadmium::make_message_bags<std::tuple<>>::type output() const{}
     TIME time_advance() const{}
 };
 
 int main(){
-    cadmium::concept::atomic_model_assert<atomic_model_with_repeated_output_ports>();
+    cadmium::concept::pdevs_atomic_model_assert<atomic_model_with_no_output_ports>();
 }
