+++ test-compile/compile-fails/pdevs_missing_time_advance_function_fails_compile_test.cpp
@@ -25,12 +25,36 @@
  */
 
 /**
- * Test that asserting coupled model with all submodels atomic does not fail compilation
+ * Test that an atomic model with no time advance function fails compilation on atomic_model_assert
  */
 
-#include "coupled_of_atomic_models.hpp"
+#include<cadmium/modeling/ports.hpp>
+#include<cadmium/concept/atomic_model_assert.hpp>
+#include<tuple>
+#include<cadmium/modeling/message_bag.hpp>
+
+/**
+ * This model has no logic, only used for structural validation tests
+ */
+template<typename TIME>
+struct atomic_model_missing_time_advance_function
+{
+    struct in : public cadmium::in_port<int>{};
+    struct out : public cadmium::out_port<int>{};
+
+    constexpr atomic_model_missing_time_advance_function() noexcept {}
+    using state_type=int;
+    state_type state=0;
+    using input_ports=std::tuple<in>;
+    using output_ports=std::tuple<out>;
+
+    void internal_transition(){}
+    void external_transition(TIME e, typename cadmium::make_message_bags<input_ports>::type mbs){}
+    void confluence_transition(TIME e, typename cadmium::make_message_bags<input_ports>::type mbs){}
+    typename cadmium::make_message_bags<output_ports>::type output() const{}
+
+};
 
 int main(){
-    cadmium::concept::coupled_model_assert<coupled_of_atomics>();
-    return 0;
+    cadmium::concept::pdevs_atomic_model_assert<atomic_model_missing_time_advance_function>();
 }
