+++ test-compile/compile-fails/pdevs_repeated_input_ports_fail_compile_test.cpp
@@ -25,7 +25,7 @@
  */
 
 /**
- * Test that an atomic model with no state fails compilation on atomic_model_assert
+ * Test that a valid atomic model does not stop compilation on atomic_model_assert.
  */
 
 #include<cadmium/modeling/ports.hpp>
@@ -34,18 +34,20 @@
 #include<cadmium/modeling/message_bag.hpp>
 
 /**
- * This model has no logic, only used for structural validation tests
+ * Test that when an atomic model has duplicated input ports, atomic_model_assert fails compilation
  */
 template<typename TIME>
-struct atomic_model_missing_state_variable
+struct atomic_model_with_repeated_input_ports
 {
-    struct in : public cadmium::in_port<int>{};
+    struct in_one : public cadmium::in_port<int>{};
+    struct in_two : public cadmium::in_port<int>{};
+
     struct out : public cadmium::out_port<int>{};
 
-    constexpr atomic_model_missing_state_variable() noexcept {}
+    constexpr atomic_model_with_repeated_input_ports() noexcept {}
     using state_type=int;
-
-    using input_ports=std::tuple<in>;
+    state_type state=0;
+    using input_ports=std::tuple<in_one, in_two, in_one>;
     using output_ports=std::tuple<out>;
 
     void internal_transition(){}
@@ -56,5 +58,5 @@ struct atomic_model_missing_state_variable
 };
 
 int main(){
-    cadmium::concept::atomic_model_assert<atomic_model_missing_state_variable>();
+    cadmium::concept::pdevs_atomic_model_assert<atomic_model_with_repeated_input_ports>();
 }
