+++ test-compile/compiles/pdevs_coupled_of_atomic_models.hpp
@@ -38,9 +38,9 @@
 #include <cadmium/concept/coupled_model_assert.hpp>
 #include <tuple>
 
-#include <cadmium/basic_model/accumulator.hpp>
-#include <cadmium/basic_model/generator.hpp>
-#include <cadmium/basic_model/passive.hpp>
+#include <cadmium/basic_model/pdevs/accumulator.hpp>
+#include <cadmium/basic_model/pdevs/generator.hpp>
+#include <cadmium/basic_model/pdevs/passive.hpp>
 
 using namespace cadmium;
 //ports
