+++ test-compile/compiles/pdevs_valid_atomic_check_test.cpp
@@ -25,7 +25,7 @@
  */
 
 /**
- * Test that an atomic model with no confluence transition fails compilation on atomic_model_assert
+ * Test that a valid atomic model does not stop compilation on atomic_model_assert.
  */
 
 #include<cadmium/modeling/ports.hpp>
@@ -37,12 +37,12 @@
  * This model has no logic, only used for structural validation tests
  */
 template<typename TIME>
-struct atomic_model_missing_confluence_function
+struct valid_atomic_model
 {
     struct in : public cadmium::in_port<int>{};
     struct out : public cadmium::out_port<int>{};
 
-    constexpr atomic_model_missing_confluence_function() noexcept {}
+    constexpr valid_atomic_model() noexcept {}
     using state_type=int;
     state_type state=0;
     using input_ports=std::tuple<in>;
@@ -50,10 +50,11 @@ struct atomic_model_missing_confluence_function
 
     void internal_transition(){}
     void external_transition(TIME e, typename cadmium::make_message_bags<input_ports>::type mbs){}
+    void confluence_transition(TIME e, typename cadmium::make_message_bags<input_ports>::type mbs){}
     typename cadmium::make_message_bags<output_ports>::type output() const{}
     TIME time_advance() const{}
 };
 
 int main(){
-    cadmium::concept::atomic_model_assert<atomic_model_missing_confluence_function>();
+    cadmium::concept::pdevs_atomic_model_assert<valid_atomic_model>();
 }
